Added a code to compute seed value without combinational logic
The previous implementation included 'while' loops which are not
synthesizable. The new implementation only uses 'for' loops which
are identical to repeating the same code several times and therefore
synthesizable.
An SVG figure is included to demonstrate the new component's
implementation.

This logic isn't finish as there are more comments by Joel that we
should address.

Also, added compilation and simulation commands to the README. This
would save us time when trying to test the code.