Adjusted 'unpacker' module to use non-combinational logic
As part of this change I added a module names 'dynamic_left_shifter'
which gets a vector and a bitmask in which a *single bit* is turned on
at most (Meaning that the bitmask can also be all zeros).
The module outputs the received vector shifted left the amount of bits
till the first 1 in the bitmask (included).

For example,
for vector v = 110010 bitmask = 001000
the module would output the vector 010000 (three shifts left).

I also added a scheme of the seed_lookup and the dynamic_left_shifter
modules in various formats.